// src/shared/api/generated/__swagger_client/models/WarehouseCreateRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WarehouseCreateRequest = {
    organizationId: string;
    orgUnitId?: string;
    code: string;
    name: string;
    address?: string;
    description?: string;
    active?: boolean;
};